Fix a bug re: entering a decimal after equals
Example prior erroneous behavior:

-Enter first number
-Select operator
-Enter second number
-Select equals
-Select decimal
-Select operator

Previously, selecting the decimal in the sequence above did not clear the
expression, so when selecting the operator in the last step, it would
incorrectly complete a calculation. This has been corrected so that
selecting a decimal with the equals button activated will clear the
expression.