Fix rendering issue & remove unnecessary code
When rendered to a mobile or Safari browser, there was a light gray line
appearing in between the 'display-small-text' container and the
'display-text' container. Adding a margin-bottom of -1px to the small text
container and a margin-top of -1px to the display container corrects this.

Also removed a line of unnecessary code